Restructure the store SPI into a core plus optional capabilities
Fold the interfaces that only existed for symmetry back into the
contracts that own them, and move the optional reporting methods off the
mandatory one. A store implementor now sees a small required core and a
handful of clearly optional capabilities instead of one large bundle.

Interface changes:
- Merge BatchMetricsStore into BatchStore. A store that runs batch
  fan-out also records its metrics, so they belong together.
- Merge ExecutionStore and JobLogStore into one JobAuditStore. Both
  record execution history.
- Add a capability(Class) probe to the composed JobStore so the engine
  can check for an optional capability at runtime instead of assuming
  every store implements every interface.
- Move the aggregate counts, rate stats, and percentile queries into a
  new JobAnalyticsStore. countPendingJobs and countActiveNodes stay on
  the core JobCrudStore because the dynamic-heartbeat loop reads them.
- Split the tag operations. Writes (insertTags, deleteTagsByJobId) stay
  on the core TagStore. Tag id lookup moves to JobQueryStore and the
  per-tag counts move to JobAnalyticsStore.

The composed JobStore still extends every interface, so the three
shipped stores keep implementing everything and behavior does not change.
The store classes keep their existing operation-class layout; the
analytics aggregator delegates the per-tag counts to the tag operations.

Re-bucket the store TCK contracts to match. Verified against MySQL,
PostgreSQL, and MongoDB.
