+++ ratchet-store-core/src/main/java/run/ratchet/store/spi/JobCrudStore.java
@@ -15,20 +15,15 @@
  */
 package run.ratchet.store.spi;
 
-import java.time.Instant;
-import java.util.EnumMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
 import run.ratchet.api.Incubating;
-import run.ratchet.api.JobPriority;
 import run.ratchet.api.JobStatus;
 import run.ratchet.api.Nullable;
 import run.ratchet.store.entity.JobEntity;
-import run.ratchet.store.entity.JobExecutionType;
 
-/** Core CRUD operations and query methods for jobs. */
+/** Core CRUD and primary-key lookup operations for jobs. */
 @Incubating
 public interface JobCrudStore {
 
@@ -101,157 +96,19 @@ public interface JobCrudStore {
    */
   List<JobEntity> findDependants(UUID parentJobId, int limit, int offset);
 
-  /** Counts pending jobs. Transaction attribute: {@code SUPPORTS}. */
+  /**
+   * Counts pending jobs. Consulted by the engine's dynamic-heartbeat backpressure loop, so it stays
+   * on the core contract rather than the optional analytics capability.
+   *
+   * <p>Transaction attribute: {@code SUPPORTS}.
+   */
   long countPendingJobs();
 
-  /** Counts jobs at the supplied status. Transaction attribute: {@code SUPPORTS}. */
-  long countJobsByStatus(JobStatus status);
-
   /**
-   * Counts jobs grouped by status.
+   * Counts currently registered scheduler nodes. Consulted by the engine's dynamic-heartbeat loop,
+   * so it stays on the core contract rather than the optional analytics capability.
    *
    * <p>Transaction attribute: {@code SUPPORTS}.
-   *
-   * <p><b>Performance note:</b> the default implementation issues one {@link
-   * #countJobsByStatus(JobStatus)} call per {@link JobStatus} constant (currently 8+). Production
-   * store implementations MUST override this with a single grouped query (e.g. {@code SELECT
-   * status, COUNT(*) FROM scheduler_job GROUP BY status}) to avoid N database round-trips on every
-   * metrics collection cycle.
    */
-  default Map<JobStatus, Long> countJobsByStatuses() {
-    Map<JobStatus, Long> counts = new EnumMap<>(JobStatus.class);
-    for (JobStatus status : JobStatus.values()) {
-      long count = countJobsByStatus(status);
-      if (count > 0) {
-        counts.put(status, count);
-      }
-    }
-    return counts;
-  }
-
-  /** Counts active jobs of the supplied type. Transaction attribute: {@code SUPPORTS}. */
-  long countActiveJobs(JobExecutionType jobType);
-
-  /** Counts currently registered scheduler nodes. Transaction attribute: {@code SUPPORTS}. */
   long countActiveNodes();
-
-  /**
-   * Counts jobs ready to execute at or before the supplied instant. Transaction attribute: {@code
-   * SUPPORTS}.
-   */
-  long countReadyJobs(Instant now);
-
-  /**
-   * Counts running jobs whose pickup timestamp is older than the supplied threshold.
-   *
-   * <p>Transaction attribute: {@code SUPPORTS}.
-   */
-  long countStuckJobs(Instant stuckThreshold);
-
-  /**
-   * Counts running jobs whose execution start time is older than the supplied threshold.
-   *
-   * <p>Transaction attribute: {@code SUPPORTS}.
-   */
-  long countLongRunningJobs(Instant threshold);
-
-  /** Counts pending batch-child jobs. Transaction attribute: {@code SUPPORTS}. */
-  long countPendingBatchChildren();
-
-  /** Counts pending jobs at the supplied priority. Transaction attribute: {@code SUPPORTS}. */
-  long countPendingJobsByPriority(JobPriority priority);
-
-  /**
-   * Counts pending jobs grouped by priority.
-   *
-   * <p>Transaction attribute: {@code SUPPORTS}.
-   *
-   * <p><b>Performance note:</b> the default implementation issues one {@link
-   * #countPendingJobsByPriority(JobPriority)} call per {@link JobPriority} constant. Production
-   * store implementations MUST override this with a single grouped query to avoid N database
-   * round-trips on every metrics collection cycle.
-   */
-  default Map<JobPriority, Long> countPendingJobsByPriorities() {
-    Map<JobPriority, Long> counts = new EnumMap<>(JobPriority.class);
-    for (JobPriority priority : JobPriority.values()) {
-      long count = countPendingJobsByPriority(priority);
-      if (count > 0) {
-        counts.put(priority, count);
-      }
-    }
-    return counts;
-  }
-
-  /** Counts pending jobs of the supplied type. Transaction attribute: {@code SUPPORTS}. */
-  long countPendingJobsByType(JobExecutionType jobType);
-
-  /**
-   * Counts pending jobs grouped by internal execution type.
-   *
-   * <p>Transaction attribute: {@code SUPPORTS}.
-   *
-   * <p><b>Performance note:</b> the default implementation issues one {@link
-   * #countPendingJobsByType(JobExecutionType)} call per {@link JobExecutionType} constant.
-   * Production store implementations MUST override this with a single grouped query to avoid N
-   * database round-trips on every metrics collection cycle.
-   */
-  default Map<JobExecutionType, Long> countPendingJobsByTypes() {
-    Map<JobExecutionType, Long> counts = new EnumMap<>(JobExecutionType.class);
-    for (JobExecutionType jobType : JobExecutionType.values()) {
-      long count = countPendingJobsByType(jobType);
-      if (count > 0) {
-        counts.put(jobType, count);
-      }
-    }
-    return counts;
-  }
-
-  /**
-   * Counts jobs in a status whose last update was at or after the supplied instant.
-   *
-   * <p>Transaction attribute: {@code SUPPORTS}.
-   */
-  long countJobsByStatusSince(JobStatus status, Instant since);
-
-  /**
-   * Counts jobs that have recorded at least one retry attempt. Transaction attribute: {@code
-   * SUPPORTS}.
-   */
-  long countJobsWithRetries();
-
-  /**
-   * Returns the fraction of recently updated jobs that have retried at least once.
-   *
-   * <p>Transaction attribute: {@code SUPPORTS}.
-   */
-  double getRetryRateStats(Instant since);
-
-  /**
-   * Returns average execution duration for jobs included in the store's metric definition.
-   *
-   * <p>Transaction attribute: {@code SUPPORTS}.
-   */
-  double getAverageProcessingTime(Instant since);
-
-  /**
-   * Returns the average number of child items for batches updated since the cutoff.
-   *
-   * <p>Transaction attribute: {@code SUPPORTS}.
-   */
-  double getAverageBatchSize(Instant since);
-
-  /**
-   * Returns the scheduled time of the oldest pending job. Transaction attribute: {@code SUPPORTS}.
-   */
-  Optional<Instant> getOldestPendingJobTime();
-
-  /**
-   * Returns the queue wait time at the given percentile for succeeded jobs.
-   *
-   * <p>Transaction attribute: {@code SUPPORTS}.
-   *
-   * @param percentile a fraction in the range [0.0, 1.0], e.g. 0.95 for p95
-   * @return queue wait time in milliseconds at the requested percentile, or 0 if no data
-   */
-  long getQueueWaitTimePercentile(double percentile);
 }
