+++ ratchet-store-core/src/main/java/run/ratchet/store/spi/TagStore.java
@@ -16,12 +16,17 @@
 package run.ratchet.store.spi;
 
 import java.util.List;
-import java.util.Map;
 import java.util.UUID;
 import run.ratchet.api.Incubating;
-import run.ratchet.api.JobStatus;
 
-/** Job tag management operations. */
+/**
+ * Job tag write operations.
+ *
+ * <p>Tag <em>writes</em> sit on the core submit path ({@code insertTags} runs during job creation),
+ * so they stay on the mandatory contract. Tag <em>reads</em> are split off: id lookup lives on
+ * {@link JobQueryStore} and the per-tag aggregate counts live on {@link JobAnalyticsStore}, both
+ * optional reporting capabilities.
+ */
 @Incubating
 public interface TagStore {
 
@@ -40,48 +45,4 @@ public interface TagStore {
    * @return number of tag rows deleted
    */
   int deleteTagsByJobId(UUID jobId);
-
-  /**
-   * Finds job ids for a tag. Transaction attribute: {@code SUPPORTS}.
-   *
-   * @param tag tag name to look up; never {@code null} or blank
-   * @param limit maximum number of ids to return; must be positive
-   * @param offset zero-based pagination offset; must be non-negative
-   * @return ordered job ids carrying the tag, never {@code null}
-   */
-  List<UUID> findJobIdsByTag(String tag, int limit, int offset);
-
-  /**
-   * Counts jobs for one tag by status. Transaction attribute: {@code SUPPORTS}.
-   *
-   * <p>This is a low-cardinality aggregation; implementations should group in the store.
-   *
-   * @param tag tag name to count against; never {@code null} or blank
-   * @return per-status counts (omitted statuses have zero matches); never {@code null}
-   */
-  Map<JobStatus, Long> countJobsByStatusForTag(String tag);
-
-  /**
-   * Counts jobs for one tag by parameter value. Transaction attribute: {@code SUPPORTS}.
-   *
-   * <p>Callers should use this for bounded diagnostic cardinalities, not arbitrary high-cardinality
-   * payload fields.
-   *
-   * @param tag tag name to count against; never {@code null} or blank
-   * @param paramKey job-parameter key whose distinct values become the result map keys; never
-   *     {@code null} or blank
-   * @return distinct-value counts keyed by parameter value (null parameter values map to {@code
-   *     null}); never {@code null}
-   */
-  Map<String, Long> countJobsByParamForTag(String tag, String paramKey);
-
-  /**
-   * Counts jobs for one tag by execution node. Transaction attribute: {@code SUPPORTS}.
-   *
-   * <p>This is bounded by scheduler-node cardinality.
-   *
-   * @param tag tag name to count against; never {@code null} or blank
-   * @return per-node counts keyed by node id; never {@code null}
-   */
-  Map<String, Long> countJobsByExecutionNodeForTag(String tag);
 }
