+++ ratchet-store-mongodb/src/test/java/run/ratchet/store/mongodb/MongoJobAuditStoreContractTest.java
@@ -13,15 +13,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package run.ratchet.store.mysql;
+package run.ratchet.store.mongodb;
 
+import org.junit.jupiter.api.AfterAll;
 import run.ratchet.store.entity.JobEntity;
 import run.ratchet.store.spi.JobStore;
-import run.ratchet.tck.store.AbstractExecutionStoreContract;
+import run.ratchet.tck.store.AbstractJobAuditStoreContract;
 
-class MysqlExecutionStoreContractTest extends AbstractExecutionStoreContract {
+/** MongoDB contract test for {@code JobAuditStore} operations. */
+class MongoJobAuditStoreContractTest extends AbstractJobAuditStoreContract {
 
-  private final MysqlTestFixture fixture = new MysqlTestFixture();
+  private static final MongoTestFixture fixture = new MongoTestFixture();
+
+  @AfterAll
+  static void closeFixture() {
+    fixture.close();
+  }
 
   @Override
   public JobStore store() {
