+++ ratchet-store-postgresql/src/main/java/run/ratchet/store/postgresql/PostgresqlAuxiliaryOperations.java
@@ -31,18 +31,13 @@ import run.ratchet.store.entity.ResourceLimitEntity;
 import run.ratchet.store.entity.WorkflowConditionEntity;
 import run.ratchet.store.id.UuidV7Factory;
 import run.ratchet.store.spi.DlqAlertStore;
-import run.ratchet.store.spi.ExecutionStore;
-import run.ratchet.store.spi.JobLogStore;
+import run.ratchet.store.spi.JobAuditStore;
 import run.ratchet.store.spi.ResourcePermitStore;
 import run.ratchet.store.spi.WorkflowConditionStore;
 import run.ratchet.store.util.RowValues;
 
 final class PostgresqlAuxiliaryOperations
-    implements ExecutionStore,
-        JobLogStore,
-        WorkflowConditionStore,
-        DlqAlertStore,
-        ResourcePermitStore {
+    implements JobAuditStore, WorkflowConditionStore, DlqAlertStore, ResourcePermitStore {
 
   private final PostgresqlStoreContext ctx;
